+++ php/modules/album.php
@@ -213,19 +213,15 @@ function setAlbumDescription($albumID, $description) {
 
 }
 
-function deleteAlbum($albumID, $delAll) {
+function deleteAlbum($albumID) {
 
 	global $database;
 
-    if ($delAll=="true") {
-        $result = $database->query("SELECT id FROM lychee_photos WHERE album = '$albumID';");
-        $error = false;
-        while($row =  $result->fetch_object()) {
-            if (!deletePhoto($row->id)) $error = true;
-        }
-    } else {
-        $result = $database->query("UPDATE lychee_photos SET album = '0' WHERE album = '$albumID';");
-        if (!$result) return false;
+	$error = false;
+
+    $result = $database->query("SELECT id FROM lychee_photos WHERE album = '$albumID';");
+    while($row =  $result->fetch_object()) {
+        if (!deletePhoto($row->id)) $error = true;
     }
 
     if ($albumID!=0) {