+++ docs/md/Changelog.md
@@ -1,11 +1,25 @@
-### v2.0
+## v2.0.1
+
+Released January 24, 2014
+
+- Share > Direct Link
+- Download individual images (Issue #43)
+- ContextMenu stays within the window (Issue #41)
+- Prevent default ContextMenu (Issue #45)
+- Small ContextMenu improvements
+- Small security improvements
+
+## v2.0
+
+Released January 22, 2014
+
 - All new redefined interface
 - Faster animations and transitions
 - Import from Dropbox
 - Import from Server
 - Download public albums
 - Several sorting options
-- Installation assistent
+- Installation assistant
 - Infobox and description for albums
 - Faster loading and improved performance
 - Better file handling and upload